Audio-Diarization/requirements.txt
streamlit==1.32.0
openai-whisper==20231117
pyannote.audio==3.1.1
torch==2.2.0
torchaudio==2.2.0
ffmpeg-python==0.2.0
pydub==0.25.1
numpy==1.26.4
matplotlib==3.8.3
plotly==5.19.0
python-dotenv==1.0.1 